+++ src/proxy/mod.rs
@@ -4,7 +4,7 @@ mod helpers;
 mod http;
 
 pub use core::*;
-pub(crate) use helpers::format_byte_count;
+pub(crate) use helpers::{clear_dns_cache, format_byte_count};
 
 #[cfg(test)]
 #[path = "tests.rs"]